requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==20.3.0
autogpt==0.0.1.dev0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bidict==0.22.1
black==23.3.0
bleach==6.1.0
blinker==1.6.2
cachelib==0.10.2
cachetools==5.3.0
certifi==2023.7.22
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cors==1.0.1
coverage==7.2.3
cryptography==41.0.1
cssselect==1.2.0
distlib==0.3.6
dnspython==2.3.0
docker==6.0.1
duckduckgo-search==2.8.6
email_validator==2.0.0.
exceptiongroup==1.1.1
filelock==3.12.4
flake8==6.0.0
Flask>0.9
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Session
Flask_Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
frozenlist==1.3.3
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
greenlet==3.0.0
gTTS==2.3.1
h11==0.14.0
httplib2==0.22.0
identify==2.5.22
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
jellyfish==0.11.2
Jinja2==3.1.2
lib50==3.0.6
loguru==0.7.0
lxml==4.9.2
MarkupSafe==2.1.3
mccabe==0.7.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nodeenv==1.7.0
numpy==1.24.2
openai==0.27.2
orjson==3.8.10
outcome==1.2.0
packaging==23.1
pathspec==0.11.1
pexpect==4.8.0
pi==0.1.2
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.2.0
playsound==1.2.2
pluggy==1.0.0
pre-commit==3.2.2
protobuf==4.22.3
pathlib == 1.0.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pyparsing==3.0.9
psycopg2==2.9.9
PySocks==1.7.1
pytest==7.3.1
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.7.1
python-socketio==5.9.0
PyYAML==5.4.1
readability-lxml==0.8.1
redis==4.5.4
regex==2023.3.23
requests==2.31.0
requests-file==1.5.1
rsa==4.9
selenium==4.8.3
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sourcery==1.2.0
SQLAlchemy==2.0.21
sqlparse==0.4.4
sympy==1.11.1
termcolor==1.1.0
tiktoken==0.3.3
tldextract==3.6.0
torch==2.0.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.17
virtualenv==20.21.0
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug
wsproto==1.2.0
WTForms==2.3.3
yarl==1.8.2
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
